Tests/LibWeb/Text/input/HTML/href-invalid.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const a = document.createElement("a");
        a.href = "http://foo:b/c";
        println(`href="${a.href}"`);
    });
</script>
