Tests/LibWeb/Text/expected/HTML/href-invalid.txt

href="http://foo:b/c"
